moved pages to a new folder called pages
