+++ src/pages/ClickerGame/ClickerGame.js
