+++ src/pages/TrumpCard/TrumpCard.js
